hrserve/pyproject.toml
[project]
name = "hrserve"
version = "1.0.0"
description = "Music model serving utilities for halfremembered"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "numpy",
]

[project.optional-dependencies]
# Audio utilities (soundfile, librosa)
audio = [
    "soundfile",
    "librosa",
]
# MIDI utilities
midi = [
    "midiutil",
    "pretty_midi",
]
# Orpheus model support (torch, x-transformers)
orpheus = [
    "torch",
    "x-transformers",
]
# OpenTelemetry
otel = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
]
# Full development setup
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hrserve"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as benchmarks",
]
